refactor: tests location
